+++ flutter_inappwebview/example/integration_test/service_worker_controller/main.dart
@@ -9,12 +9,14 @@ import '../util.dart';
 part 'set_service_worker_client.dart';
 part 'should_intercept_request.dart';
 part 'include_cookies_on_should_intercept_request.dart';
+part 'client_survives_second_controller.dart';
 
 void main() {
   final shouldSkip = !ServiceWorkerController.isClassSupported();
 
   skippableGroup('Service Worker Controller', () {
     shouldInterceptRequest();
+    clientSurvivesSecondController();
     setServiceWorkerClient();
     serviceWorkerIncludeCookies();
   }, skip: shouldSkip);